+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ 0.9.1 ] - 2024-12-26
+## [ 0.9.1 ] - 2025-01-01
+### Added
+- Added existFile, simplifyPath, removeFile, removeDirectory, writeFileContent in FileUtil.
+
+### Changed
+- Optimized trimLeft and trimRight in StringUtil.
 
 ## [ 0.9.0 ] - 2024-12-26
 ### Added